# proof of concept: use video scaling instead of NPB luminance