NPB luminance setting dims dark colors to black, this gives really ugly results at low brightness:
Examples:
- palette effect with rainbow is just red, green, blue even with black stripes 
- DNA Spiral: parts of the spiral are missing at brightness < 10

The proposed method applies video scaling: any color that is non zero will stay non zero
NPB luminance is not used anymore.

fixes https://github.com/wled/WLED/issues/4794