@DedeHai 2025-07-31
@blazoncek I am not sure this is the proper/best way to implement it, feel free to move stuff around.