@blazoncek 2025-07-31
It is not. At least not with `NeoPixelBusLg`.

If you intend to replace luminance (and since 0.16 with segment blending you can) use base class `NeoPixelBus` instead. And it should be done in `show()` which outputs segment data to LEDs.

The other way is to move gamma handling into `NeoPixelBusLg` as it is designed with that in mind. I did intend to do that at some point (it is not very good practice to introduce too many changes at once).