@DedeHai 2025-07-31
I thought so :)
Also this draft does not preserve color ratios well, some palettes get ugly hue distortion at low brightness. 
How can I test if NPB handles gamma better? That would indeed be the best way.