@DedeHai 2025-07-31
I am not going to mess with NPB, I want to focus on other things so I will leave that for someone else. I will update this PR with an improved solution that can be used in the meantime.