@willmmiles 2025-07-31
> @blazoncek I am a bit lost on how to properly implement this as I don't understand all the busses i.e. digitalbus vs polybus and need some guidance: where to set the brightness if not using luminance?

I think the key question is, "do we get any benefits from letting different bus types use custom luminance or gamma calculations"?   Ie. should this be the responsibility of the bus, or should we have a single formula we always use?

IF we want to use a common formula for all buses, the calculation code belongs in [`WS2812FX::show()`](https://github.com/wled/WLED/blob/b76ef231fcb97073285488dff0dadfdcd2ace188/wled00/FX_fcn.cpp#L1657) before we pass the pixel color to the bus -- same as we're doing for gamma correction.   This will give us consistent results across all output types.  If we do this, we can replace `NeoPixelBusLg<>` with plain `NeoPixelBus<>` and maybe save some cycles and code space.   

IF instead we think there's some advantage to letting the bus decide how to apply the brightness calculation, then we should integrate the new calculation (and possibly move gamma correction) directly in to `NeoPixelBusLg`, which exists specifically to apply a luminance and gamma adjustment to pixel colors on the way through to the hardware buffer.