@willmmiles 2025-07-31
> I think the key question is, "do we get any benefits from letting different bus types use custom luminance or gamma calculations"? Ie. should this be the responsibility of the bus, or should we have a single formula we always use?

At first glance I'd push in favor of using a common formula.  It might be that there are special cases (network outputs?) where passing a global brightness separately is preferable -- do we know of any?