@blazoncek 2025-07-31
@willmmiles is right. IIRC at some point @Makuna considered abandoning NeoPixelBrightnessBus (or whatever it was called) but ultimately replaced it with NeoPixelBusLg.

At the time it was impossible to use NeoPixelBus (non Lg or Brightness) as entire WLED logic relied on NeoPixelBus to do the brightness scaling and WLED also handled gamma in an inappropriate way. Both these issues are resolved and indeed brightness scaling and/or gamma correction can be done in show() (and not in a bus).

We can stick with Lg however and let it do brightness/gamma for digital LEDs. For analog we will need to add gamma adjustment to bus as it already have brightness. For network/virtual bus we'd also need to do the same as it expects gamma to be applied.