@willmmiles 2025-07-31
> We can stick with Lg however and let it do brightness/gamma for digital LEDs. For analog we will need to add gamma adjustment to bus as it already have brightness. For network/virtual bus we'd also need to do the same as it expects gamma to be applied.

I did note that `NeoPixelBusLg` does brightness before gamma correction.   Right now we're doing it the other way around (gamma, then brightness).

I'll also add that, using this branch and doing the brightness ourselves for digital busses, switching from `NeoPixelBusLg` to `NeoPixelBus` saves 3.5kb of code space (if that still matters).

To confirm, then, you think we should go the other way around, ie.push the gamma calculation down so each bus type does it?    NeoPixelBusLg bakes its brightness algorithm in fairly deeply -- to adopt the new one here, we can't really use NeoPixelBusLg either way.

One up side of doing the brightness and gamma in the bus object is that we could later support different gamma values for each bus -- could be helpful if you're tweaking mismatched hardware or something.