@blazoncek 2025-08-01
> I did note that `NeoPixelBusLg` does brightness before gamma correction.

That would be the correct way.

> switching from `NeoPixelBusLg` to `NeoPixelBus` saves 3.5kb of code space (if that still matters)

I think it does and IMO worth doing (as it will also simplify bus wrapper).

> you think we should go the other way around, ie.push the gamma calculation down so each bus type does it?

Not yet 100% sure but, to answer your next question, we don't need to use Lg, we can implement brightness and gamma in bus manager (`Bus` class and derived classes). It is fairly easy implementation and as you say it allows us to have different brightness scaling and/or gamma for each bus.