@DedeHai 2025-08-01
I ran a test with applying brightness before gamma: it looks awful.