@DedeHai 2025-08-01
I did see the comment in the code about `NeoGammaWLEDMethod` and if the `NeoGammaTableMethod` is not fundamentally different, which I assume its not as the output looks very similar, using gamma and brightness from NPBLg is a no-go.
I will update the code of this PR with a new suggestion in show().