@DedeHai 2025-08-01
replacing NPB was easier than I thought, did that in the last commit.
@TripleWhy [the last commit](https://github.com/wled/WLED/pull/4798/commits/eb2de8a57fb95684b4469922300717bcd151bbcf) is probably as good as it gets with regards to color preservation at low brightness.