@DedeHai 2025-08-02
Ran some speed comparison:
This PR gives me 2% higher FPS in my scenario (ESP C3, 64x32 DDP pixels at 40FPS)

I deem this PR read to merge if this is the route we want to go, which I highly recommend for quality reasons, see https://github.com/wled/WLED/issues/4794

@netmindz what is your opinion on this? 
TL;DR: 
- low brightness color accuracy is bad 
- using NPBlg (currently used) allows for gamma correction, but if used, it gets even worse
- Ditching the lg variant of NPB saves some flash (I tested 1.3k on C3, @willmmiles tested >3.5k)
- This PR fixes the issue at hand but the question is, do we want to fix it some other way (i.e. per bus and not globally)