@DedeHai 2025-08-02
thanks for the feedback, the "not turning off LEDs" part is a good find, I did not check that part of code logic.

> It doesn't appear to produce even brightness at least until around brightness 30, above it, I can't tell. It looks much more even to me without gamma correction. Even brightness is the point of gamma correction, isn't it?

I do not know what you mean by that. Gamma is to convert linear color space to perceived color space. if you mean linear brightness increase, then you need to enable "apply gamma to brightness" which should still work the same as before.