@DedeHai 2025-08-02
it has no obligation to look "even" (please try to use more technical terms, that can mean anything).
why we bother:
add two colors colors, say one is B=60, the other one is R=120. dim that down by 50%  you get B=30 R=60. Now convert that to perceived brightness of each channel and the hue changes. this is what gamma correction fixes. Do a side-by side comparison of a palette, say orangery with and without gamma correction at different brightness and you will see it.