@DedeHai 2025-08-03
I had an idea:
since many are used to and seem to like the old way of applying brightness and 0.16 is a breaking change in that regard: 
adding a checkmark in settings to make "apply gamma after Effects" optional, if not set, gamma is set to default 2.2 and only applied to palettes. Downside is that if someone wants to only apply gamma to brightness the gamma value is fixed to 2.2.