@DedeHai 2025-08-26
Finally finished the re-work of the bus brightness handling and the ABL. Please review. 
See [commit comment](https://github.com/wled/WLED/pull/4798/commits/00062b2460ae024e7f7bc89b6295ae076ca7b1a9) for details.

Hope no changes got lost in manual merging of my WIP branch. I reworked the code several times over to arrive at this version which IMHO is the best compromise between speed and code consistency. Global ABL would be faster if done on strip level, only scaling once but it convolutes code and adds computational overhead even if the limit does not apply. @blazoncek took that approach in his dev branch (see table)

I did some testing on the ESP32 C3 but not extensively. Here is a summary of my findings:

Flash usage: only checked for C3 and ESP32
- C3: saves 1.1k
- ESP32: saves 5.7k

Test with 2x 256 LEDs (2 outputs) on C3, 32x16 matrix, palette effect

| Version   | Limiter Disabled | Global Limit (non-limiting / limiting) | Per-Bus Limit (non-limiting / limiting) |
|-----------|------------------|----------------------------------------|------------------------------------------|
| PR        | 76.3 FPS         | 75.6 FPS (-0.7) / 70.0 FPS (-6.3)     | 74.8 FPS (-1.5) / 69.3 FPS (-6.3)        |
| Main      | 75.0 FPS         | 74.1 FPS (-0.9) / 74.0 FPS (-0.9)     | 68.0 FPS (-6.1) / 61.3 FPS (-12.8)       |
| Blaz dev  | 67.0 FPS         | 64.0 FPS (-3.0) / 64.0 FPS (-3.0)     | 63.5 FPS (-3.5) / 60.0 FPS (-7.0)        |

In summary:
- this PR fixes low brightness color handling for much nicer colors
- moved brightness scaling from NPB into WLED code, saves us some flash and gives a slight speed boost
- complete rework of ABL, now also respects white channel and mapping and calculations are more consistent. Also it is done on bus level only, no more ABL code on strip level